requirements.txt
asgiref==3.5.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-admin-honeypot==1.1.0
# django-baton==2.4.1
django-crispy-forms==1.14.0
django-grappelli==3.0.3
django-hijack==3.2.1
django-import-export==2.9.0
django-jet-reboot==1.3.2
et-xmlfile==1.1.0
feedparser==6.0.10
google-api-python-client==1.4.1
google-auth==2.12.0
google-auth-httplib2==0.1.0
httplib2==0.20.4
idna==3.4
MarkupPy==1.14
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.10
# psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
waitress==2.1.1
xlrd==2.0.1
xlwt==1.3.0
